+++ packages/common/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @powersync/common
 
+## 1.33.2
+
+### Patch Changes
+
+- 9b2bde3: Fix compilation error due to broken import in `SyncProgress.d.ts`
+
+## 1.33.1
+
+### Patch Changes
+
+- ffe3095: Improve websocket keepalive logic to reduce keepalive errors.
+- 36d8f28: Remove unused compaction logic.
+- 53236a8: Rust client: Properly upload CRUD entries made while offline.
+- b7255b7: Rust sync client: Fix `connect()` resolving before a connection is made.
+- 70a9cf5: Added CommonJs output for common package. Ensuring default export entry is last so that require statements don't use .mjs output instead of .cjs.
+- d1b7fcb: Rust sync client: Fix reported `lastSyncedAt` values in sync status.
+
 ## 1.33.0
 
 ### Minor Changes